Incremental VK builder test for all ring suites
