+++ src/suites/bandersnatch.rs
@@ -145,60 +145,4 @@ pub(crate) mod tests {
             .map(|p| p.check(true).ok())
             .is_some());
     }
-
-    #[cfg(feature = "ring")]
-    #[test]
-    pub fn ring_verifier_key_builder() {
-        use crate::ring::{testing::RingSuiteExt, Prover, Verifier};
-        use crate::testing::{random_val, random_vec};
-
-        let mut rng = ark_std::test_rng();
-        let ring_ctx = ThisSuite::context();
-
-        let sk = Secret::from_seed(b"");
-        let pk = sk.public();
-
-        let ring_size = ring_ctx.max_ring_size();
-        let prover_idx = random_val::<usize>(Some(&mut rng)) % ring_size;
-        let mut pks = random_vec::<AffinePoint>(ring_size, Some(&mut rng));
-        pks[prover_idx] = pk.0;
-
-        let input = Input::from(random_val(Some(&mut rng)));
-        let output = sk.output(input);
-
-        let prover_key = ring_ctx.prover_key(&pks);
-        let prover = ring_ctx.prover(prover_key, prover_idx);
-        let proof = sk.prove(input, output, b"foo", &prover);
-
-        // Incremental ring verifier key construction
-        let (mut vk_builder, loader) = ring_ctx.verifier_key_builder();
-        assert_eq!(vk_builder.free_slots(), pks.len());
-
-        let extra_pk = random_val::<AffinePoint>(Some(&mut rng));
-        let faulty_loader = |_| None;
-        assert_eq!(
-            vk_builder.append(&[extra_pk], faulty_loader).unwrap_err(),
-            usize::MAX
-        );
-
-        loop {
-            let chunk_len = 1 + random_val::<usize>(Some(&mut rng)) % 5;
-            let chunk = pks.drain(..pks.len().min(chunk_len)).collect::<Vec<_>>();
-            if chunk.is_empty() {
-                break;
-            }
-            println!("Appending {} items", chunk.len());
-            vk_builder.append(&chunk[..], &loader).unwrap();
-            assert_eq!(vk_builder.free_slots(), pks.len());
-        }
-        // No more space left
-        let extra_pk = random_val::<AffinePoint>(Some(&mut rng));
-        assert_eq!(vk_builder.append(&[extra_pk], &loader).unwrap_err(), 0);
-
-        let verifier_key = vk_builder.finalize();
-
-        let verifier = ring_ctx.verifier(verifier_key);
-        let result = Public::verify(input, output, b"foo", &proof, &verifier);
-        assert!(result.is_ok());
-    }
 }
